# Optimize DateTime decomposition in CoreLib