--- a/src/libraries/System.Private.CoreLib/src/System/Globalization/EastAsianLunisolarCalendar.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Globalization/EastAsianLunisolarCalendar.cs
@@ -400,7 +400,8 @@ namespace System.Globalization
         private DateTime LunarToTime(DateTime time, int year, int month, int day)
         {
             LunarToGregorian(year, month, day, out int gy, out int gm, out int gd);
-            return GregorianCalendar.GetDefaultInstance().ToDateTime(gy, gm, gd, time.Hour, time.Minute, time.Second, time.Millisecond);
+            time.GetTime(out int hour, out int minute, out int second, out int millisecond);
+            return GregorianCalendar.GetDefaultInstance().ToDateTime(gy, gm, gd, hour, minute, second, millisecond);
         }
 
         private void TimeToLunar(DateTime time, out int year, out int month, out int day)