@ts2do 2020-01-21
Based on what I'm seeing in additional runs and taking into account the nature of the changes, the 1-3% variance in either direction is noise.