@danmosemsft 2020-01-21
Thanks for this optimization!